Cleanup
Added fallback options for the taskbar module in case the config is not found or is disabled.
Made the text length an unsigned int to prevent the compiler from complaining.
Used unparent instead of hide to "attempt to reduce resource usage" (Ideally this should be done in the module widget itself not here)
Added some margins to the text widget.