Switch popovers to an overlay
Working with popovers was probably the most painful part of this bar.
Too many issues honestly, Not worth the effort.
Using an overlay window *should* work better.

At the moment i lazily switched out the popovers in favor of a general overlay window, A lot of work still needs to be done here..