(EXPERIMENTAL) Use grid layout
Use a grid layout to allow for more customization options.
At the moment this is the same as before.
Unsure what would be better, Give the end user full control which could cause a widget to not show if the user inputs something incorrectly, Or let the dingus handle everything on it's own.