Add window tracking
Added window tracking, This is sort of a way to customize the behavior of the bar based on what the current focused window is.

At the moment this is limited to CSS only but in the future i'm hoping to reuse this to block notifications if you're in a game for example.

Also fixed some stuff in the config.